Move constants and helper functions to separate files
